Modul deneme VM-V1/EdgeSolution/modules/AnalyticModule/requirements.txt
azure-iot-device
pyModbusTCP
schedule

